# api-gateway/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid input."
          },
          "409": {
            "description": "Conflict - Email already in use."
          }
        },
        "summary": "Register a new user",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful, returns cookies.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          },
          "401": {
            "description": "Unauthorized - Invalid credentials."
          }
        },
        "summary": "User login",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Token successfully refreshed."
          },
          "401": {
            "description": "Unauthorized - Invalid or expired refresh token."
          }
        },
        "summary": "Refresh access token using refresh token cookie",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Profile retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logout successful."
          },
          "401": {
            "description": "Unauthorized."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Logout and clear cookies",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/email-templates": {
      "post": {
        "operationId": "EmailTemplatesController_create",
        "parameters": [
          {
            "name": "x-workspace-id",
            "in": "header",
            "description": "Active workspace ID for resource isolation",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmailTemplateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "email-templates"
        ]
      },
      "get": {
        "operationId": "EmailTemplatesController_findAll",
        "parameters": [
          {
            "name": "x-workspace-id",
            "in": "header",
            "description": "Active workspace ID for resource isolation",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based).",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size.",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Filter by name fragment.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "email-templates"
        ]
      }
    },
    "/email-templates/{id}": {
      "get": {
        "operationId": "EmailTemplatesController_findOne",
        "parameters": [
          {
            "name": "x-workspace-id",
            "in": "header",
            "description": "Active workspace ID for resource isolation",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "email-templates"
        ]
      },
      "patch": {
        "operationId": "EmailTemplatesController_update",
        "parameters": [
          {
            "name": "x-workspace-id",
            "in": "header",
            "description": "Active workspace ID for resource isolation",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmailTemplateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "email-templates"
        ]
      },
      "delete": {
        "operationId": "EmailTemplatesController_remove",
        "parameters": [
          {
            "name": "x-workspace-id",
            "in": "header",
            "description": "Active workspace ID for resource isolation",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "email-templates"
        ]
      }
    }
  },
  "info": {
    "title": "TalentFlow AI API",
    "description": "API documentation for the TalentFlow AI API Gateway",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your access token"
      }
    },
    "schemas": {
...
      "SignupDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "The email address of the user"
          },
          "password": {
            "type": "string",
            "example": "StrongPass123!",
            "description": "The password for the account (min 8 chars, 1 uppercase, 1 lowercase, 1 number, 1 special char)",
            "minLength": 8
          },
          "fullName": {
            "type": "string",
            "example": "John Doe",
            "description": "The full name of the user"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "RECRUITER",
              "INTERVIEWER"
            ],
            "example": "RECRUITER",
            "description": "The role of the user"
          }
        },
        "required": [
          "email",
          "password",
          "fullName",
          "role"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "description": "The unique identifier of the user"
          },
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "The email address of the user"
          },
          "fullName": {
            "type": "string",
            "example": "John Doe",
            "description": "The full name of the user"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "RECRUITER",
              "INTERVIEWER"
            ],
            "example": "RECRUITER",
            "description": "The role of the user"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2023-01-01T00:00:00Z",
            "description": "The date and time when the user was created"
          }
        },
        "required": [
          "id",
          "email",
          "fullName",
          "role",
          "createdAt"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "description": "The user details",
            "allOf": [
              {
                "$ref": "#/components/schemas/AuthResponseDto"
              }
            ]
          },
          "message": {
            "type": "string",
            "example": "Login successful",
            "description": "A message indicating the result of the login operation"
          }
        },
        "required": [
          "user",
          "message"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "The email address of the user"
          },
          "password": {
            "type": "string",
            "example": "StrongPass123!",
            "description": "The password for the account"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
      "CreateEmailTemplateDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The template name (unique within the workspace).",
            "example": "Interview Invitation"
          },
          "subject": {
            "type": "string",
            "description": "The subject line of the email template.",
            "example": "You are invited to interview"
          },
          "body": {
            "type": "string",
            "description": "The body of the email template (plain text or Handlebars).",
            "example": "Hello {{candidateName}}, ..."
          }
        },
        "required": [
          "name",
          "subject",
          "body"
        ]
      },
      "UpdateEmailTemplateDto": {
        "type": "object",
        "properties": {
          "subject": {
            "type": "string",
            "description": "Updated subject line."
          },
          "body": {
            "type": "string",
            "description": "Updated body."
          }
        }
      }
    }
  }
}